Added verification of LF structure
LF structure is verified by computing sequence length for each
sequence and summing. Not guaranteed to be accurate on a
successful output, but likely. This is currently much slower than
the rest of the program, it is O(n) LF operation. Everything else is
O(r). 5 minutes on mtb152.fmd. Estimated time on human472:
1000 minutes. Needs to be parallelized to speed up. Verification will
not be done in the future, it is mainly there for testing purposes
during implementation.